Update README.md
This is great. I think we have a great handle of what we are offering to people who buy in to the campaign:
1. Offer them inside info and tours of snowboarding trails
2. First looks to try out gear and GoPro their experiences
3. One on one training with a pro snowboarder

The above are the rewards essentially, but what we should specify is what end we are doing this for. For instance, will the funds pay for a snowboarding tutorial website we will be launching? Or perhaps the funds will go toward preserving/maintaining snowboarding trails?